requirements.txt
google-generativeai>=0.3.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
googleapis-common-protos>=1.56.0
google-api-python-client>=2.88.0
gspread>=5.10.0
pillow>=10.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
aiofiles>=23.0.0
httpx>=0.24.0
pdf2image>=1.16.3
pypdf2>=3.0.1